+++ CHANGELOG.md
@@ -1,4 +1,5 @@
 # Changelog
+
 ## Release 1.0.0
 * Initial Release